refactor: constants, utils 파일명 변경
