+++ src/common/Card.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled, { css } from 'styled-components';
-import { CARD, COLOR, FONT_SIZE, FONT_WEIGHT } from '../constants/constants';
+import { CARD, COLOR, FONT_SIZE, FONT_WEIGHT } from '../constants/constant';
 
 const AddCard = css`
   .add {
