+++ src/components/ModalContents/CardColor/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import { CARD } from '../../../constants/constants';
+import { CARD } from '../../../constants/constant';
 import CardColorItem from './CardColorItem';
 
 const CardColorWrapper = styled.div`
