+++ src/components/NewCardForm/CardCVCInput/index.styles.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { COLOR, FONT_SIZE } from '../../../constants/constants';
+import { COLOR, FONT_SIZE } from '../../../constants/constant';
 
 export const CardCVCInputWrapper = styled.div`
   margin-bottom: 10px;
